Add triggerable C/C++ debugging
This PR allows triggering a debugger on the C++ land from the target
language.
This can be very helpful during development and debugging.

Two macros are provided:
- GDB()
- GDB_ON_TRIGGER()

The first triggers gdb/ddd. The second is the same but only triggers
in case the method `trigger_gdb()` in `utils/gdb.h` was called before.

These macros only generate code when LGBM is compiled with
`-DUSE_DEBUG=ON`,
otherwise they add up no overhead.

Every method in the C API now starts with a `GDB_ON_TRIGGER()`,
meaning
all methods in the C API can be debugged promptly.

The debugger will kick in and the program will wait until
`signal_gdb_attached()` is called by the user.
This will then unlock and resume the program execution.
In case the debugger fails to kick in, it's easy to attach with `gdb
--pid=<target_pid>`, and the program will periodically print the
target pid until `signal_gdb_attached()` is called.